* add: impostor command

* change: changed colour from “yellow” to “red” in remember.js

* change: changed colour from “darkpurple” to “red” in remembercheck.js

* add: log code to index.js

* add: logSchema

* add: log setup and disable

* fix: admin ghost ping msg

* fix: eco schema not storing data

* fix: /rob comand

* fix: undefined in /beg

* add: imagine image command

* add: fake-tweet command

* change: changed openai api key in chatgpt

* changed: changed openai api key in image generator

* change: openai api config

* change: openai api config

* change: changed latency ping

* add: added new dependencies to package.json

* add: added new dependencies to package-lock.json

* add: add coolkids.json

* add: add status command (cool kids only)

* remove: coolkids.json

* change: changed handling of defined cool kids

* Remove: commands directory for restructure

* add: Re added commands under new file structure :)

* remove: openai api config1

* remove: openai config2

+++ src/index.js
@@ -490,6 +490,10 @@ client.on(Events.GuildMemberAdd, async member => {
 
         if (number == 2) time = 60;
         if (number >= 3) time = 500;
+        
+        if (message.member.permissions.has(PermissionsBitField.Flags.Administrator)) {
+            return;
+        }
 
         const msg = await message.channel.send({ content: `${message.author}, Dont try to ghost ping please.`});
         setTimeout(() => msg.delete(), 5000);
